Code Cleanup
Do some cleanup around Float vs Double, and some capitalization.

+++ Sources/PWM/Expansion.swift
@@ -84,7 +84,7 @@ class ExpansionPWM : Module, CustomStringConvertible {
 
 class ExpansionPWMChannel : Channel {
     let token: String
-    var luminance: Float {
+    var luminance: Double {
         didSet { self.onLuminanceChanged() }
     }
